fix: address CI clippy + leaf duplicates + Bedrock leaf persistence
- biome.rs: drop redundant `as u64` cast (c is already u64; -D warnings
  in CI flagged clippy::unnecessary_cast).
- tree.rs: remove duplicate `(0, _, -1)` column from SPRUCE_LEAVES_FILL_*
  and PINE_LEAVES_FILL_* (carried over from the original SPRUCE_LEAVES_FILL
  pattern). The duplicate produced the same leaves twice; trimming to 5
  entries matches the oak/birch shape and removes wasted writes.
- bedrock_block_map.rs: add explicit persistent_bit + update_bit states
  for mangrove_leaves and azalea_leaves (matching the cherry_leaves
  pattern) so Bedrock exports don't decay the new leaves.

cargo clippy --all-targets --all-features -- -D warnings clean.

+++ src/biome.rs
@@ -152,7 +152,7 @@ mod tests {
         }
         let longs = pack_biome_indices(&indices, 1);
         assert_eq!(longs.len(), 1);
-        let expected: u64 = (0..64u64).fold(0, |acc, c| acc | (((c % 2) as u64) << c));
+        let expected: u64 = (0..64u64).fold(0, |acc, c| acc | ((c % 2) << c));
         assert_eq!(longs[0] as u64, expected);
     }
